Update AdminPanelProvider.php
Spatie's documentation explicitly states that an array of parameters can be used in place of a file :
```
Path to the client secret json file. Take a look at the README of this package to learn how to get this file. You can also pass the credentials as an array instead of a file path.
```